dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev-requirements.in
#
alabaster==0.7.12
    # via sphinx
appdirs==1.4.4
    # via
    #   black
    #   virtualenv
arrow==0.16.0
    # via jinja2-time
attrs==19.3.0
    # via
    #   black
    #   pytest
autopep8==1.5.4
    # via -r dev-requirements.in
babel==2.8.0
    # via sphinx
bcrypt==3.2.0
    # via paramiko
binaryornot==0.4.4
    # via cookiecutter
black==19.10b0
    # via -r dev-requirements.in
cerberus==1.3.2
    # via molecule
certifi==2020.6.20
    # via requests
cffi==1.14.3
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
chardet==3.0.4
    # via
    #   binaryornot
    #   requests
click-completion==0.5.2
    # via molecule
click-help-colors==0.8
    # via molecule
click==7.1.2
    # via
    #   black
    #   click-completion
    #   click-help-colors
    #   cookiecutter
    #   molecule
colorama==0.4.3
    # via rich
commonmark==0.9.1
    # via rich
cookiecutter==1.7.2
    # via molecule
coverage==5.2.1
    # via pytest-cov
cryptography==3.1.1
    # via paramiko
dataclasses==0.8
    # via
    #   molecule
    #   rich
distlib==0.3.1
    # via virtualenv
distro==1.5.0
    # via selinux
docker==4.4.1
    # via
    #   -r dev-requirements.in
    #   molecule-docker
docutils==0.16
    # via sphinx
enrich==1.2.6
    # via molecule
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
flake8-polyfill==1.0.2
    # via pep8-naming
flake8==3.8.3
    # via
    #   -r dev-requirements.in
    #   flake8-polyfill
idna==2.10
    # via requests
imagesize==1.2.0
    # via sphinx
importlib-metadata==1.7.0
    # via
    #   flake8
    #   pluggy
    #   pytest
    #   tox
    #   virtualenv
importlib-resources==3.0.0
    # via virtualenv
isort==5.5.4
    # via -r dev-requirements.in
jinja2-time==0.2.0
    # via cookiecutter
jinja2==2.11.2
    # via
    #   click-completion
    #   cookiecutter
    #   jinja2-time
    #   molecule
    #   sphinx
markupsafe==1.1.1
    # via
    #   cookiecutter
    #   jinja2
mccabe==0.6.1
    # via flake8
molecule-docker==0.2.4
    # via -r dev-requirements.in
molecule==3.2.2
    # via
    #   -r dev-requirements.in
    #   molecule-docker
more-itertools==8.4.0
    # via pytest
mypy-extensions==0.4.3
    # via mypy
mypy==0.782
    # via -r dev-requirements.in
packaging==20.4
    # via
    #   molecule
    #   pytest
    #   sphinx
    #   tox
paramiko==2.7.2
    # via molecule
pathspec==0.8.0
    # via
    #   black
    #   yamllint
pep8-naming==0.11.1
    # via -r dev-requirements.in
pluggy==0.13.1
    # via
    #   molecule
    #   pytest
    #   tox
poyo==0.5.0
    # via cookiecutter
py==1.9.0
    # via
    #   pytest
    #   tox
pycodestyle==2.6.0
    # via
    #   autopep8
    #   flake8
pycparser==2.20
    # via cffi
pyflakes==2.2.0
    # via flake8
pygments==2.6.1
    # via
    #   rich
    #   sphinx
pynacl==1.4.0
    # via paramiko
pyparsing==2.4.7
    # via packaging
pytest-cov==2.10.0
    # via -r dev-requirements.in
pytest==5.4.3
    # via
    #   -r dev-requirements.in
    #   pytest-cov
    #   testinfra
python-dateutil==2.8.1
    # via arrow
python-slugify==4.0.1
    # via cookiecutter
pytz==2020.1
    # via babel
pyyaml==5.3.1
    # via
    #   molecule
    #   yamllint
regex==2020.7.14
    # via black
requests==2.24.0
    # via
    #   cookiecutter
    #   docker
    #   sphinx
rich==9.8.2
    # via
    #   enrich
    #   molecule
selinux==0.2.1
    # via
    #   molecule
    #   molecule-docker
shellingham==1.3.2
    # via click-completion
six==1.15.0
    # via
    #   bcrypt
    #   click-completion
    #   cookiecutter
    #   cryptography
    #   docker
    #   packaging
    #   pynacl
    #   python-dateutil
    #   tox
    #   virtualenv
    #   websocket-client
snowballstemmer==2.0.0
    # via sphinx
sphinx==3.1.2
    # via -r dev-requirements.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==1.0.3
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.4
    # via sphinx
subprocess-tee==0.2.0
    # via molecule
testinfra==5.3.1
    # via -r dev-requirements.in
text-unidecode==1.3
    # via python-slugify
toml==0.10.1
    # via
    #   autopep8
    #   black
    #   tox
tox==3.18.0
    # via -r dev-requirements.in
typed-ast==1.4.1
    # via
    #   black
    #   mypy
typing-extensions==3.7.4.2
    # via
    #   mypy
    #   rich
urllib3==1.25.10
    # via requests
virtualenv==20.0.28
    # via tox
wcwidth==0.2.5
    # via pytest
websocket-client==0.57.0
    # via docker
yamllint==1.25.0
    # via molecule
zipp==3.1.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
